requirements.txt
torch
scikit-learn
pandas
faiss-gpu
datasets
gdown
ragatouille
colbert-ai